test: Seclude resolveUseCasesConfig to outer module
